[#5] feat: 계좌서비스 인터페이스, 구현체 분리
