+++ src/main/java/com/example/hanaharmonybackend/web/controller/AccountController.java
@@ -6,7 +6,6 @@ import com.example.hanaharmonybackend.service.AccountService;
 import com.example.hanaharmonybackend.util.SecurityUtil;
 import com.example.hanaharmonybackend.web.dto.account.AccountResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
